requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cleanup==9.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1

# Development and testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
